backend/requirements.txt
# InspiLarder - 灵感食仓
# 食物记录单页WebApp - 后端依赖

# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
aiomysql==0.2.0
alembic==1.12.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI与图像处理
openai==1.3.0
pillow==10.1.0
aiofiles==23.2.1

# 工具库
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3

# 开发工具（可选）
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
